Split packaging integration tests into two groups
